ai_agent_learning/langgraph/exam-agent/pyproject.toml
[project]
name = "exam-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.1",
    "langchain-groq>=1.1.2",
    "langgraph>=1.1.9",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
    "tavily-python>=0.7.23",
]
